+++ pty/pty_windows.go
@@ -37,14 +37,11 @@ func newPty(opt ...Option) (PTY, error) {
 		return nil, xerrors.Errorf("pty not supported")
 	}
 
-	// For cases where no process will be attached (e.g., ptytest.New() used with
-	// Attach() for in-process CLI testing), we use a pipe-based implementation
-	// that doesn't require ConPTY. ConPTY requires an attached process to function
-	// correctly - without one, the pipe handles become invalid intermittently.
-	//
-	// When Start() is called, it creates its own ConPTY with the process properly
-	// attached, so this pipe-based PTY is only for the Attach() use case.
-	return newTestPTY()
+	// On Windows, pty.New() without Start() is only used by ptytest.New() for
+	// in-process CLI testing. ConPTY requires an attached process to function
+	// correctly, so ptytest has its own pipe-based implementation. Production
+	// code should use pty.Start() which creates a ConPTY with process attached.
+	return nil, xerrors.Errorf("pty without process not supported on Windows; use ptytest.New() for tests")
 }
 
 // newConPty creates a PTY backed by a Windows PseudoConsole (ConPTY). This
@@ -99,92 +96,6 @@ func newConPty(opt ...Option) (*ptyWindows, error) {
 	}
 
 	return pty, nil
-}
-
-// testPTY is a pipe-based PTY implementation for in-process CLI testing on
-// Windows. It's used by ptytest.New() when no real process will be attached.
-type testPTY struct {
-	// The pipe for "input" - what would be stdin for a process.
-	// Test writes to inputWriter, CLI reads from inputReader.
-	inputReader *os.File
-	inputWriter *os.File
-
-	// The pipe for "output" - what would be stdout for a process.
-	// CLI writes to outputWriter, test reads from outputReader.
-	outputReader *os.File
-	outputWriter *os.File
-
-	closeMutex sync.Mutex
-	closed     bool
-}
-
-func newTestPTY() (*testPTY, error) {
-	p := &testPTY{}
-
-	var err error
-	p.inputReader, p.inputWriter, err = os.Pipe()
-	if err != nil {
-		return nil, xerrors.Errorf("create input pipe: %w", err)
-	}
-	p.outputReader, p.outputWriter, err = os.Pipe()
-	if err != nil {
-		_ = p.inputReader.Close()
-		_ = p.inputWriter.Close()
-		return nil, xerrors.Errorf("create output pipe: %w", err)
-	}
-
-	return p, nil
-}
-
-func (*testPTY) Name() string {
-	return ""
-}
-
-// Input returns a ReadWriter for the input pipe. The test writes to this
-// (simulating user input), and the CLI reads from it.
-func (p *testPTY) Input() ReadWriter {
-	return ReadWriter{
-		Reader: p.inputReader,
-		Writer: p.inputWriter,
-	}
-}
-
-// Output returns a ReadWriter for the output pipe. The CLI writes to this,
-// and the test reads from it to verify output.
-func (p *testPTY) Output() ReadWriter {
-	return ReadWriter{
-		Reader: p.outputReader,
-		Writer: p.outputWriter,
-	}
-}
-
-func (*testPTY) Resize(uint16, uint16) error {
-	// Resize is a no-op for pipe-based PTY since there's no real terminal.
-	return nil
-}
-
-func (p *testPTY) Close() error {
-	p.closeMutex.Lock()
-	defer p.closeMutex.Unlock()
-	if p.closed {
-		return nil
-	}
-	p.closed = true
-
-	var firstErr error
-	if err := p.outputWriter.Close(); err != nil && firstErr == nil {
-		firstErr = err
-	}
-	if err := p.outputReader.Close(); err != nil && firstErr == nil {
-		firstErr = err
-	}
-	if err := p.inputWriter.Close(); err != nil && firstErr == nil {
-		firstErr = err
-	}
-	if err := p.inputReader.Close(); err != nil && firstErr == nil {
-		firstErr = err
-	}
-	return firstErr
 }
 
 type ptyWindows struct {
