+++ pty/ptytest/ptytest_other.go
@@ -1,0 +1,9 @@
+//go:build !windows
+
+package ptytest
+
+import "github.com/coder/coder/v2/pty"
+
+func newTestPTY(opts ...pty.Option) (pty.PTY, error) {
+	return pty.New(opts...)
+}
